Initial commit of revised website.
This initial commit makes it easier for anyone to run their own version of the SM Elite stats website. It includes a data pack (NO PACK) and instructions on how to build a custom themed data pack.

There are minor CSS/spacing issues for some columns due to the increased width of the content.